VoiceAgentsOpenaiTest/requirements.txt
openai>=1.40.0
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0
python-dotenv>=1.0.1
rich>=13.7.0
flask>=3.0.0
flask-cors>=4.0.0
